+++ src/test/kotlin/tornadofx/testapps/InternalWindowTest.kt
@@ -41,6 +41,6 @@ class Editor : View("Editor") {
 
     private fun save() {
         alert(CONFIRMATION, "Saved!", "You did it!")
-        closeModal()
+        close()
     }
 }